+++ src/main/java/hwr/oop/board/ChessBoardException.java
@@ -1,4 +1,4 @@
-package hwr.oop;
+package hwr.oop.board;
 
 public class ChessBoardException extends Exception {
   public ChessBoardException(String message) {
